Menambahkan fungsi hapus pada halaman input,  menambahkan download dan memperbaiki beberapa hal

+++ skp.php
@@ -75,8 +75,6 @@ if (isset($_POST["inputskp"])) {
             <div class="card shadow mb-4">
                 <div class="card-header py-3">
                     <h5 class="font-weight-bold text-primary text-center">SKP tahun <?= date('Y'); ?></h5>
-                    <i class="fas fa-info fa-sm fa-fw mr-2 text-gray-760" data-toggle="tooltip" data-placement="top" title="Reload Halaman Jika Anda Menambahkan Data Baru !">
-                    </i>
                 </div>
                 <div class="card-body">
 
@@ -85,9 +83,8 @@ if (isset($_POST["inputskp"])) {
                             <tr>
                                 <th scope="col">#</th>
                                 <th scope="col">Deskripsi</th>
-                                <th scope="col">Target</th>
-                                <th scope="col">Satuan</th>
-                                <th scope="col">Edit</th>
+                                <th scope="col" width="115">Target</th>
+                                <th scope="col" width="71">Edit</th>
                             </tr>
                         </thead>
                         <tbody><?php $i = 1; ?>
@@ -95,10 +92,8 @@ if (isset($_POST["inputskp"])) {
                                 <tr>
                                     <th scope="row"><?= $i; ?></th>
                                     <td><?= $row['uraian']; ?></td>
-                                    <td><?= $row['target']; ?></td>
-                                    <td><?= $row['satuan']; ?></td>
-                                    <td> <a href="ubah_skp.php?id=<?= $row['id_skp']; ?>" class="fas fa-pen fa-sm fa-fw mr-1"> </a>|
-                                        <a href="hapus_skp.php?id=<?= $row['id_skp']; ?>" class="fas fa-trash fa-sm fa-fw mr-1"> </a></td>
+                                    <td><?= $row['target'] . ' ' . $row['satuan']; ?></td>
+                                    <td> <a href="ubah_skp.php?id=<?= $row['id_skp']; ?>" class="fas fa-pen fa-sm fa-fw mr-1"> </a>|<a href="hapus_skp.php?id=<?= $row['id_skp']; ?>" onclick="return confirm('Yakin Ingin Hapus data ?');" class="fas fa-trash fa-sm fa-fw mr-1"> </a></td>
                                 </tr>
                                 <?php $i++; ?>
                             <?php endforeach; ?>
@@ -140,6 +135,62 @@ if (isset($_POST["inputskp"])) {
 include 'template/modal.php';
 ?>
 
+<?php
+// $data = query("SELECT * FROM skp WHERE id_skp = $id")[0];
+
+if (isset($_POST["ubahskp"])) {
+
+    if (ubahSkp($_POST) > 0) {
+        echo "<script>
+                alert('Data Berhasil Di Ubah !');
+                document.location.href= 'skp.php';
+            </script>";
+    } else {
+        echo mysqli_error($koneksi);
+    }
+}
+
+?>
+<!-- Modal -->
+<div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h5 class="modal-title" id="exampleModalLabel">Ubah Data SKP</h5>
+                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+            </div>
+            <div class="modal-body">
+                <input type="text" name="data" id="data" value="">
+                <form action="" method="POST">
+                    <input type="hidden" name="id" value="<?= $data['id_skp']; ?>">
+                    <div class="form-group">
+                        <label for="exampleFormControlTextarea1">Sasaran Kerja</label>
+                        <textarea name="uraian" class="form-control" id="exampleFormControlTextarea1" rows="2"><?= $data['uraian']; ?></textarea>
+                    </div>
+                    <div class="form-group col-lg-6">
+                        <label for="exampleFormControlInput1">Target Kerja</label>
+                        <input type="number" name="target" value="<?= $data['target']; ?>" class="form-control" id="exampleFormControlInput1" placeholder="Input Angka">
+                    </div>
+                    <div class="form-group">
+                        <label for="exampleFormControlInput1">Satuan</label>
+                        <input type="text" name="satuan" value="<?= $data['satuan']; ?>" class="form-control" id="exampleFormControlInput1" placeholder="Satuan dari target (misal : dokumen atau gambar)">
+                    </div>
+
+                    <button type="submit" name="ubahskp" class="btn btn-primary">Ubah Data</button>
+
+                </form>
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                <button type="button" class="btn btn-primary">Save changes</button>
+            </div>
+        </div>
+    </div>
+</div>
+
+
 <!-- Bootstrap core JavaScript-->
 <script src="vendor/jquery/jquery.min.js"></script>
 <script src="vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
@@ -151,6 +202,24 @@ include 'template/modal.php';
 <script src="js/sb-admin-2.min.js"></script>
 <script>
     $('[data-toggle="tooltip"]').tooltip();
+</script>
+<script>
+    $(document).ready(function() {
+
+        $('.getid').click(function() {
+            var c_id = $('.getid').data('id');
+            //var c_id= $(this).data(id) - checked both
+            $("#data").val(c_id);
+            // var data_id = '';
+
+            // if (typeof $(this).data('id') !== 'undefined') {
+
+            //     data_id = $(this).data('id');
+            // }
+
+            // $('#my_element_id').val(data_id);
+        })
+    });
 </script>
 
 </body>
